+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 
 - [x] feat(client): choose scenario (Fri 2)
       Allows to load scenarios from a directory.
+- [x] feat(client): updates' preferences for DPO (Wed 7)
 
 ### July 2024